Introduce ValidatedFilePath to separate file path validation from composition
