+++ tests/Unit/Path/FilePathTest.php
@@ -7,7 +7,6 @@ namespace Haspadar\Piqule\Tests\Unit\Path;
 use Haspadar\Piqule\File\FileName;
 use Haspadar\Piqule\Path\DirectoryPath;
 use Haspadar\Piqule\Path\FilePath;
-use Haspadar\Piqule\PiquleException;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
@@ -24,16 +23,5 @@ final class FilePathTest extends TestCase
             ))->value(),
             'Composes absolute file path from directory path and file name',
         );
-    }
-
-    #[Test]
-    public function throwsWhenNameEndsWithSeparator(): void
-    {
-        $this->expectException(PiquleException::class);
-
-        (new FilePath(
-            new DirectoryPath('/var/www'),
-            new FileName('dir/'),
-        ))->value();
     }
 }
